requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
mysql-connector-python==9.1.0
